Fix GLM4 dtype issues and add memory-constrained testing
- Fix RoPE cos/sin dtype to match input tensor dtype
- Fix model.to() to include dtype conversion
- Add --max-layers option for testing on memory-constrained GPUs
- Verified working: 10.88 tok/s decode with 10 layers on 16GB GPU

+++ v2/model/glm4.py
@@ -76,8 +76,8 @@ class PartialRotaryEmbedding(nn.Module):
         x: torch.Tensor,
         position_ids: torch.Tensor,
     ) -> Tuple[torch.Tensor, torch.Tensor]:
-        cos = self.cos_cached[position_ids]
-        sin = self.sin_cached[position_ids]
+        cos = self.cos_cached[position_ids].to(x.dtype)
+        sin = self.sin_cached[position_ids].to(x.dtype)
         return cos.unsqueeze(1), sin.unsqueeze(1)
